createNotFoundPageWithNoIdontHave
