+++ jiseong-choi/.eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\jiseo\\Desktop\\react-project1\\jiseong-choi\\src\\index.js":"1","C:\\Users\\jiseo\\Desktop\\react-project1\\jiseong-choi\\src\\App.js":"2","C:\\Users\\jiseo\\Desktop\\react-project1\\jiseong-choi\\src\\Components\\Views\\NotFoundPage\\NotFoundPage.js":"3","C:\\Users\\jiseo\\Desktop\\react-project1\\jiseong-choi\\src\\components\\views\\NotFoundPage\\NotFoundPage.js":"4","C:\\Users\\jiseo\\Desktop\\react-project1\\jiseong-choi\\src\\components\\views\\LandingPage\\LandingPage.js":"5","C:\\Users\\jiseo\\Desktop\\react-project1\\jiseong-choi\\src\\components\\views\\RegisterPage\\RegisterPage.js":"6","C:\\Users\\jiseo\\Desktop\\react-project1\\jiseong-choi\\src\\components\\views\\LoginPage\\LoginPage.js":"7","C:\\Users\\jiseo\\Desktop\\react-project1\\jiseong-choi\\src\\components\\views\\Header\\Header.js":"8","C:\\Users\\jiseo\\Desktop\\react-project1\\jiseong-choi\\src\\components\\views\\NewLoginPage\\NewLoginPage.js":"9"},{"size":232,"mtime":1611898312886,"results":"10","hashOfConfig":"11"},{"size":989,"mtime":1612193649035,"results":"12","hashOfConfig":"11"},{"size":168,"mtime":1611896580546,"results":"13","hashOfConfig":"11"},{"size":168,"mtime":1611896580546,"results":"14","hashOfConfig":"11"},{"size":165,"mtime":1611896477057,"results":"15","hashOfConfig":"11"},{"size":2214,"mtime":1612192184166,"results":"16","hashOfConfig":"11"},{"size":1825,"mtime":1612192140451,"results":"17","hashOfConfig":"11"},{"size":252,"mtime":1612104809549,"results":"18","hashOfConfig":"11"},{"size":1853,"mtime":1612195744522,"results":"19","hashOfConfig":"11"},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},"6fyr13",{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"35","messages":"36","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"37","messages":"38","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\jiseo\\Desktop\\react-project1\\jiseong-choi\\src\\index.js",[],["39","40"],"C:\\Users\\jiseo\\Desktop\\react-project1\\jiseong-choi\\src\\App.js",[],"C:\\Users\\jiseo\\Desktop\\react-project1\\jiseong-choi\\src\\Components\\Views\\NotFoundPage\\NotFoundPage.js",[],"C:\\Users\\jiseo\\Desktop\\react-project1\\jiseong-choi\\src\\components\\views\\NotFoundPage\\NotFoundPage.js",[],"C:\\Users\\jiseo\\Desktop\\react-project1\\jiseong-choi\\src\\components\\views\\LandingPage\\LandingPage.js",[],"C:\\Users\\jiseo\\Desktop\\react-project1\\jiseong-choi\\src\\components\\views\\RegisterPage\\RegisterPage.js",[],"C:\\Users\\jiseo\\Desktop\\react-project1\\jiseong-choi\\src\\components\\views\\LoginPage\\LoginPage.js",[],"C:\\Users\\jiseo\\Desktop\\react-project1\\jiseong-choi\\src\\components\\views\\Header\\Header.js",[],"C:\\Users\\jiseo\\Desktop\\react-project1\\jiseong-choi\\src\\components\\views\\NewLoginPage\\NewLoginPage.js",[],{"ruleId":"41","replacedBy":"42"},{"ruleId":"43","replacedBy":"44"},"no-native-reassign",["45"],"no-negated-in-lhs",["46"],"no-global-assign","no-unsafe-negation"]
+[{"C:\\Users\\jiseo\\Desktop\\react-project1\\jiseong-choi\\src\\index.js":"1","C:\\Users\\jiseo\\Desktop\\react-project1\\jiseong-choi\\src\\App.js":"2","C:\\Users\\jiseo\\Desktop\\react-project1\\jiseong-choi\\src\\Components\\Views\\NotFoundPage\\NotFoundPage.js":"3","C:\\Users\\jiseo\\Desktop\\react-project1\\jiseong-choi\\src\\components\\views\\NotFoundPage\\NotFoundPage.js":"4","C:\\Users\\jiseo\\Desktop\\react-project1\\jiseong-choi\\src\\components\\views\\LandingPage\\LandingPage.js":"5","C:\\Users\\jiseo\\Desktop\\react-project1\\jiseong-choi\\src\\components\\views\\RegisterPage\\RegisterPage.js":"6","C:\\Users\\jiseo\\Desktop\\react-project1\\jiseong-choi\\src\\components\\views\\LoginPage\\LoginPage.js":"7","C:\\Users\\jiseo\\Desktop\\react-project1\\jiseong-choi\\src\\components\\views\\Header\\Header.js":"8"},{"size":232,"mtime":1611898312886,"results":"9","hashOfConfig":"10"},{"size":848,"mtime":1612283157747,"results":"11","hashOfConfig":"10"},{"size":168,"mtime":1611896580546,"results":"12","hashOfConfig":"10"},{"size":468,"mtime":1612283093072,"results":"13","hashOfConfig":"10"},{"size":165,"mtime":1611896477057,"results":"14","hashOfConfig":"10"},{"size":2290,"mtime":1612282326272,"results":"15","hashOfConfig":"10"},{"size":2074,"mtime":1612281851588,"results":"16","hashOfConfig":"10"},{"size":252,"mtime":1612104809549,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},"6fyr13",{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},"C:\\Users\\jiseo\\Desktop\\react-project1\\jiseong-choi\\src\\index.js",[],["35","36"],"C:\\Users\\jiseo\\Desktop\\react-project1\\jiseong-choi\\src\\App.js",[],"C:\\Users\\jiseo\\Desktop\\react-project1\\jiseong-choi\\src\\Components\\Views\\NotFoundPage\\NotFoundPage.js",[],"C:\\Users\\jiseo\\Desktop\\react-project1\\jiseong-choi\\src\\components\\views\\NotFoundPage\\NotFoundPage.js",["37"],"C:\\Users\\jiseo\\Desktop\\react-project1\\jiseong-choi\\src\\components\\views\\LandingPage\\LandingPage.js",[],"C:\\Users\\jiseo\\Desktop\\react-project1\\jiseong-choi\\src\\components\\views\\RegisterPage\\RegisterPage.js",[],"C:\\Users\\jiseo\\Desktop\\react-project1\\jiseong-choi\\src\\components\\views\\LoginPage\\LoginPage.js",[],"C:\\Users\\jiseo\\Desktop\\react-project1\\jiseong-choi\\src\\components\\views\\Header\\Header.js",[],{"ruleId":"38","replacedBy":"39"},{"ruleId":"40","replacedBy":"41"},{"ruleId":"42","severity":1,"message":"43","line":7,"column":13,"nodeType":"44","endLine":7,"endColumn":254},"no-native-reassign",["45"],"no-negated-in-lhs",["46"],"jsx-a11y/iframe-has-title","<iframe> elements must have a unique title property.","JSXOpeningElement","no-global-assign","no-unsafe-negation"]
