+++ jiseong-choi/src/App.js
@@ -5,7 +5,6 @@ import LoginPage from './components/views/LoginPage/LoginPage.js';
 import NotFoundPage from './components/views/NotFoundPage/NotFoundPage.js';
 import RegisterPage from './components/views/RegisterPage/RegisterPage.js';
 import Header from './components/views/Header/Header'
-import NewLoginPage from './components/views/NewLoginPage/NewLoginPage.js';
 
 function App() {
 	return (	  
@@ -15,7 +14,6 @@ function App() {
 				<Switch>					
 					<Route exact path="/" component={LandingPage} />					
 					<Route exact path="/login" component={LoginPage} />	
-					<Route exact path="/newLogin" component={NewLoginPage} />
 					<Route exact path="/register" component={RegisterPage} />					
 					<NotFoundPage />					
 				</Switch>
