--- jiseong-choi/src/components/views/NewLoginPage/NewLoginPage.css
@@ -1,63 +1,0 @@
-.login-wrapper {
-    width: 100%;
-    height: 90vh;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.login-container {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    border-radius: 4px;
-    width: 300px;
-    height: 350px;
-    margin: 45px auto;
-    background-color: #f8f8f8;
-}
-
-.login-form {
-    display: flex;
-    flex-direction: column;
-}
-
-.login-form__input {
-    margin: 15px 0;
-    width: 170px;
-    height: 30px;
-    border: none;
-    background-color: #f8f8f8;
-    border-bottom: 2px solid #d5e6e0;
-}
-
-.login-form__input::placeholder {
-    text-align: center;
-}
-
-.login-form__input:focus {
-    transition: 1s ease-in-out;
-    border-color: #35C361;
-    outline: none;
-}
-
-.login-form__title {
-    display: flex;
-    justify-content: center;
-    margin-bottom: 20px;
-    font-size: 32px;
-}
-
-.login-button {
-    margin-top: 20px;
-    height: 30px;
-    border: none;
-    background-color: #44B78B;
-    color: white;
-    border-radius: 2px;
-}
-
-.login-button:hover {
-    transition: 0.5s ease-in-out;
-    background-color: #1C8364;
-}
