+++ jiseong-choi/src/components/views/NotFoundPage/NotFoundPage.css
@@ -1,0 +1,7 @@
+.not-found__container {
+    width: 100%;
+    height: 90vh;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
