+++ jiseong-choi/src/components/views/NotFoundPage/NotFoundPage.js
@@ -1,9 +1,10 @@
 import React from 'react'
+import './NotFoundPage.css'
 
 function NotFoundPage() {
     return (
-        <div>
-            NotFoundPage
+        <div className="not-found__container">
+            <iframe width="560" height="315" src="https://www.youtube.com/embed/qZtY8Q0AXbQ?autoplay=1&mute=1" frameBorder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowFullScreen autoPlay="1"></iframe>
         </div>
     )
 }
